Create Expense tracker
🚀 Exciting Announcement! 🚀Hello LinkedIn community! 👋 I'm thrilled to share a major update from my Week 3 Python Internship at MotionCut ! 🌐✨
The task involves The Expense Tracker project  that is designed to reinforce our understanding of Python programming
concepts and enhance our skills in building practical applications. In this project, we will be
developing an Expense Tracker application that allows users to manage their expenses efficiently.
This real-world application will involve handling data, user input, and implementing key
functionalities.
After running this code we will add category ,amount and description in expense tracker and that will automatically calculate the expenditure.